MVP/Prototype Created
The most basic prototype has been created. A button controls a gate, and you want to match the color of the circles to the colors of what's available. You let Green circles go through to the end, crush Yellow circles with the gate, and block Red circles from getting passed anything.